chore: remove old posthog experiment
