+++ src/components/TeacherComponents/LessonOverviewHeaderDesktop/index.tsx
@@ -1,2 +1,1 @@
 export { LessonOverviewHeaderDesktop } from "./LessonOverviewHeaderDesktop";
-export { LessonOverviewHeaderDesktopB } from "./LessonOverviewHeaderDesktopB";
