+++ src/components/TeacherComponents/LessonOverviewHeaderMobile/index.tsx
@@ -1,2 +1,1 @@
 export { LessonOverviewHeaderMobile } from "./LessonOverviewHeaderMobile";
-export { LessonOverviewHeaderMobileB } from "./LessonOverviewHeaderMobileB";
